backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
openai-whisper==20240930
pydub==0.25.1
ollama==0.4.7
